requirements.txt
python-telegram-bot==21.5
httpx
python-dotenv==0.21.0
beautifulsoup4
urllib3